+++ config/vuetify.js
@@ -1,0 +1,49 @@
+export default {
+  customVariables: ['~/assets/variables.scss'],
+  defaultAssets: false,
+  icons: {
+    iconfont: 'mdiSvg',
+  },
+  theme: {
+    options: {
+      customProperties: true,
+    },
+    dark: false,
+    themes: {
+      light: {
+        primary: '#EF1D13',
+        elusColor: '#EEAB00',
+        culturelColor: '#80BA27',
+        sportColor: '#772056',
+        galaColor: '#0096AA',
+        solidaritesColor: '#ED6C26',
+        techniquesColor: '#B4B2B1',
+        othersColor: '#4b6584',
+        cafetColor: '#78e08f',
+        partenariatsColor: '#38ada9',
+        shopColor: '#079992',
+        alumniColor: '#a55eea',
+        documentsColor: '#eb3b5a',
+        elevesColor: '#a55eea',
+        tutorinsaColor: '#4c98d2',
+      },
+      dark: {
+        primary: '#EF1D13',
+        elusColor: '#EEAB00',
+        culturelColor: '#80BA27',
+        sportColor: '#772056',
+        galaColor: '#0096AA',
+        solidaritesColor: '#ED6C26',
+        techniquesColor: '#B4B2B1',
+        othersColor: '#4b6584',
+        cafetColor: '#78e08f',
+        partenariatsColor: '#38ada9',
+        shopColor: '#079992',
+        alumniColor: '#a55eea',
+        documentsColor: '#eb3b5a',
+        elevesColor: '#a55eea',
+        tutorinsaColor: '#4c98d2',
+      },
+    },
+  },
+}
