+++ nuxt.config.js
@@ -1,3 +1,8 @@
+import head from './config/head'
+import feed from './config/feed'
+import sitemap from './config/sitemap'
+import vuetify from './config/vuetify'
+
 export default {
   /*
    ** Nuxt rendering mode
@@ -16,48 +21,7 @@ export default {
    ** Headers of the page
    ** See https://nuxtjs.org/api/configuration-head
    */
-  head: {
-    titleTemplate: '%s' + ' - Campus INSA CVL',
-    title: 'CHARGEMENT...',
-    meta: [
-      { charset: 'utf-8' },
-      { name: 'viewport', content: 'width=device-width, initial-scale=1' },
-      {
-        hid: 'og:locale',
-        property: 'og:locale',
-        content: 'fr_FR',
-      },
-      {
-        hid: 'og:type',
-        property: 'og:type',
-        content: 'website',
-      },
-      {
-        hid: 'og:image',
-        property: 'og:image',
-        content:
-          'https://raw.githubusercontent.com/Campus-INSA-CVL/campus-website/master/static/icon.png?token=AKZLUYA3V5QXSSQYTXWCOLC7ACPPW',
-      },
-      {
-        hid: 'og:site_name',
-        property: 'og:site_name',
-        content: 'Campus INSA CVL',
-      },
-      {
-        hid: 'twitter:card',
-        name: 'twitter:card',
-        content: 'summary',
-      },
-
-      {
-        hid: 'twitter:image',
-        name: 'twitter:image',
-        content:
-          'https://raw.githubusercontent.com/Campus-INSA-CVL/campus-website/master/static/icon.png?token=AKZLUYA3V5QXSSQYTXWCOLC7ACPPW',
-      },
-    ],
-    link: [{ rel: 'icon', type: 'image/x-icon', href: '/favicon.ico' }],
-  },
+  head,
   /*
    ** Customize the progress-bar color
    */
@@ -138,130 +102,12 @@ export default {
    ** Feed module configuration
    ** See https://github.com/nuxt-community/feed-module
    */
-  feed: {
-    data: [
-      // {
-      //   path: 'culturel',
-      //   title: 'Le blog du pôle Culturel',
-      //   description: 'Mauris congue orci eget ornare scelerisque.',
-      // },
-      // {
-      //   path: 'solidarites',
-      //   title: 'Pôle Solidarités',
-      //   description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
-      // },
-    ],
-    factory: (data) => {
-      const baseUrlArticles = 'https://campus.insa-cvl.fr'
-
-      const { $content } = require('@nuxt/content')
-
-      return data.map(({ path, title, description }) => {
-        const createFeedArticles = async function (feed) {
-          feed.options = {
-            title,
-            description,
-            link: baseUrlArticles,
-          }
-
-          const articles = await $content(`federation/${path}/blog`).fetch()
-
-          articles.forEach((article) => {
-            const url = `${baseUrlArticles}/federation/${path}/blog/${article.slug}`
-
-            feed.addItem({
-              title: article.title,
-              id: url,
-              link: url,
-              date: new Date(article.updatedAt),
-              description: article.description,
-              author: article.author,
-            })
-          })
-        }
-        return {
-          path: `/feed/${path}/rss.xml`,
-          type: 'rss2',
-          create: createFeedArticles,
-        }
-      })
-    },
-  },
+  feed,
   /*
    ** Sitemap module configuration
    ** See https://github.com/nuxt-community/sitemap-module
    */
-  sitemap: {
-    hostname: 'https://campus.insa-cvl.fr',
-    gzip: true,
-    defaults: {
-      changefreq: 'weekly',
-      priority: 1,
-      lastmod: new Date(),
-    },
-    routes: async () => {
-      const { $content } = require('@nuxt/content')
-      const dynRoutes = [
-        '/vie-etudiante',
-        '/services',
-        '/outils',
-        '/federation/index',
-        '/federation/culturel/associations',
-        '/federation/elus/representation',
-        '/federation/solidarites/associations',
-        '/federation/sport/associations',
-        '/federation/sport/sports',
-        '/federation/techniques/associations',
-        // '/federation/culturel/blog',
-        // '/federation/solidarites/blog',
-      ]
-      let path = []
-
-      for await (const route of dynRoutes) {
-        if (route.includes('associations')) {
-          const content = await $content(route).fetch()
-          path = [
-            ...content.associations.map((file) => {
-              return {
-                url: file.path,
-                changefreq: 'monthly',
-                priority: 0.3,
-                lastmod: new Date(),
-              }
-            }),
-            ...path,
-          ]
-        } else if (route === '/federation/index') {
-          const content = await $content(route).fetch()
-          path = [
-            ...content.federations.map((file) => {
-              return {
-                url: file.path,
-                changefreq: 'monthly',
-                priority: 0.7,
-                lastmod: new Date(),
-              }
-            }),
-            ...path,
-          ]
-        } else {
-          const content = await $content(route).only(['path']).fetch()
-          path = [
-            ...content.map((file) => {
-              return {
-                url: file.path,
-                changefreq: 'weekly',
-                priority: 0.5,
-                lastmod: new Date(),
-              }
-            }),
-            ...path,
-          ]
-        }
-      }
-      return path
-    },
-  },
+  sitemap,
   /*
    ** Robots module configuration
    ** See https://github.com/nuxt-community/robots-module
@@ -274,55 +120,7 @@ export default {
    ** vuetify module configuration
    ** https://github.com/nuxt-community/vuetify-module
    */
-  vuetify: {
-    customVariables: ['~/assets/variables.scss'],
-    defaultAssets: false,
-    icons: {
-      iconfont: 'mdiSvg',
-    },
-    theme: {
-      options: {
-        customProperties: true,
-      },
-      dark: false,
-      themes: {
-        light: {
-          primary: '#EF1D13',
-          elusColor: '#EEAB00',
-          culturelColor: '#80BA27',
-          sportColor: '#772056',
-          galaColor: '#0096AA',
-          solidaritesColor: '#ED6C26',
-          techniquesColor: '#B4B2B1',
-          othersColor: '#4b6584',
-          cafetColor: '#78e08f',
-          partenariatsColor: '#38ada9',
-          shopColor: '#079992',
-          alumniColor: '#a55eea',
-          documentsColor: '#eb3b5a',
-          elevesColor: '#a55eea',
-          tutorinsaColor: '#4c98d2',
-        },
-        dark: {
-          primary: '#EF1D13',
-          elusColor: '#EEAB00',
-          culturelColor: '#80BA27',
-          sportColor: '#772056',
-          galaColor: '#0096AA',
-          solidaritesColor: '#ED6C26',
-          techniquesColor: '#B4B2B1',
-          othersColor: '#4b6584',
-          cafetColor: '#78e08f',
-          partenariatsColor: '#38ada9',
-          shopColor: '#079992',
-          alumniColor: '#a55eea',
-          documentsColor: '#eb3b5a',
-          elevesColor: '#a55eea',
-          tutorinsaColor: '#4c98d2',
-        },
-      },
-    },
-  },
+  vuetify,
   /*
    ** Webfontloader module configuration
    ** https://github.com/Developmint/nuxt-webfontloader
